fixes to make source compile as C++ (tested with vc9)

+++ trunk/src/lib3ds_quat.c
@@ -187,7 +187,7 @@ lib3ds_quat_ln(float c[4]) {
     double om, s, t;
 
     s = sqrt(c[0] * c[0] + c[1] * c[1] + c[2] * c[2]);
-    om = atan2(s, c[3]);
+    om = atan2(s, (double)c[3]);
     if (fabs(s) < LIB3DS_EPSILON) {
         t = 0.0f;
     } else {
